delete old admin page, refactor api structure, delete images on delete doc, fix create and edit with missing params, add toast for missing required fields
