--- src/components/admin/AdminLayout.tsx
@@ -1,80 +1,0 @@
-import React from "react";
-import { useAuth } from "../../providers/auth";
-import { toast } from "react-toastify";
-
-interface AdminLayoutProps {
-  children: React.ReactNode;
-  onCreateClick: () => void;
-  createButtonText: string;
-}
-
-export const AdminLayout: React.FC<AdminLayoutProps> = ({
-  children,
-  onCreateClick,
-  createButtonText,
-}) => {
-  const { user, logout } = useAuth();
-
-  const handleLogout = async () => {
-    try {
-      await logout();
-      toast.success("Successfully logged out!");
-    } catch {
-      toast.error("Failed to log out");
-    }
-  };
-
-  return (
-    <div className="min-h-screen bg-gray-50">
-      {/* Header */}
-      <div className="bg-white shadow-sm border-b">
-        <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-          <div className="flex justify-between items-center h-16">
-            <div className="flex items-center">
-              <h1 className="text-2xl font-bold text-gray-900">Admin Panel</h1>
-            </div>
-
-            <div className="flex items-center space-x-4">
-              {createButtonText && (
-                <button
-                  onClick={onCreateClick}
-                  className="inline-flex items-center px-4 py-2 border border-transparent text-sm font-medium rounded-md text-white bg-blue-600 hover:bg-blue-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-blue-500"
-                >
-                  <svg
-                    className="w-4 h-4 mr-2"
-                    fill="none"
-                    stroke="currentColor"
-                    viewBox="0 0 24 24"
-                  >
-                    <path
-                      strokeLinecap="round"
-                      strokeLinejoin="round"
-                      strokeWidth={2}
-                      d="M12 4v16m8-8H4"
-                    />
-                  </svg>
-                  {createButtonText}
-                </button>
-              )}
-
-              <div className="flex items-center space-x-3">
-                <span className="text-sm text-gray-500">{user?.email}</span>
-                <button
-                  onClick={handleLogout}
-                  className="inline-flex items-center px-3 py-2 border border-gray-300 shadow-sm text-sm font-medium rounded-md text-gray-700 bg-white hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-blue-500"
-                >
-                  Logout
-                </button>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-
-      {/* Main Content */}
-      <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 py-8">
-        {children}
-      </div>
-    </div>
-  );
-};
